rename Dto to QueryResult
